refactor:홈화면조회 api 수정
