+++ src/main/java/com/deare/backend/api/home/service/HomeService.java
@@ -1,10 +1,10 @@
-package com.deare.backend.domain.home.service;
+package com.deare.backend.api.home.service;
 
-import com.deare.backend.domain.home.dto.HomeDashboardResponse;
-import com.deare.backend.domain.home.dto.HomeSettingDto;
-import com.deare.backend.domain.home.dto.HomeStickerDto;
-import com.deare.backend.domain.home.dto.HomeUserDto;
-import com.deare.backend.domain.home.exception.HomeErrorCode;
+import com.deare.backend.api.home.dto.HomeDashboardResponse;
+import com.deare.backend.api.home.dto.HomeSettingDto;
+import com.deare.backend.api.home.dto.HomeStickerDto;
+import com.deare.backend.api.home.dto.HomeUserDto;
+import com.deare.backend.api.home.exception.HomeErrorCode;
 import com.deare.backend.domain.setting.entity.UserSetting;
 import com.deare.backend.domain.setting.repository.UserSettingRepository;
 import com.deare.backend.domain.sticker.entity.UserSticker;
@@ -14,6 +14,7 @@ import com.deare.backend.domain.user.repository.UserRepository;
 import com.deare.backend.global.common.exception.GeneralException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -25,6 +26,7 @@ public class HomeService {
     private final UserSettingRepository userSettingRepository;
     private final UserStickerRepository stickerRepository;
 
+    @Transactional(readOnly=true)
     public HomeDashboardResponse getHome(Long userId) {
 
         User user = userRepository.findById(userId)
@@ -68,11 +70,11 @@ public class HomeService {
                 sticker.getId(),
                 sticker.getImage().getId(),
                 sticker.getImage().getImageUrl(),
-                sticker.getPosX().doubleValue(),
-                sticker.getPosY().doubleValue(),
+                sticker.getPosX(),
+                sticker.getPosY(),
                 sticker.getPosZ(),
-                sticker.getRotation().doubleValue(),
-                sticker.getScale().doubleValue()
+                sticker.getRotation(),
+                sticker.getScale()
         );
     }
 }
